requirements.txt
python-telegram-bot==20.8
instagrapi==1.19.8
Pillow==10.3.0
python-dotenv==1.0.1
aiohttp==3.9.1
pytesseract==0.3.10
opencv-python==4.12.0.88
httpx==0.25.2
vk-api==11.9.9
requests==2.32.5
